mcp-template/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "mcp-template"
version = "0.1.0"
description = "Reusable Model Context Protocol server template"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "InboxGreen", email = "dev@inboxgreen.local" }
]
dependencies = [
  "typer>=0.12,<0.13",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
  "rich>=13,<14"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.8",
  "mypy>=1.11"
]

[project.scripts]
"mcp-template" = "mcp_template.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "W"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_template"]

[tool.hatch.build.targets.sdist]
include = [
  "src/mcp_template",
  "scripts",
  "tools",
  "docker",
  "docs",
  "config",
  "tasks.json",
  "MCP_TEMPLATE_BLUEPRINT.md"
]
